+++ internal/config/config.go
@@ -55,10 +55,11 @@ func NewConfig() *Config {
 	c := &Config{
 		// Agent defaults:
 		Agent: &AgentConfig{
-			Interval:      internal.Duration{Duration: 10 * time.Second},
-			RoundInterval: true,
-			FlushInterval: internal.Duration{Duration: 10 * time.Second},
-			FlushJitter:   internal.Duration{Duration: 5 * time.Second},
+			Interval:              internal.Duration{Duration: 10 * time.Second},
+			RoundInterval:         true,
+			FlushInterval:         internal.Duration{Duration: 10 * time.Second},
+			FlushJitter:           internal.Duration{Duration: 5 * time.Second},
+			WebhookServiceAddress: ":1925",
 		},
 
 		Tags:          make(map[string]string),
@@ -122,6 +123,9 @@ type AgentConfig struct {
 	Quiet        bool
 	Hostname     string
 	OmitHostname bool
+
+	// Address for listening webhooks
+	WebhookServiceAddress string
 }
 
 // Inputs returns a list of strings of the configured inputs.
@@ -219,6 +223,9 @@ var header = `# Telegraf Configuration
   ## If set to true, do no set the "host" tag in the telegraf agent.
   omit_hostname = false
 
+  ## Address to listen for webhook
+  webhook_service_address = ":1925"
+
 
 ###############################################################################
 #                            OUTPUT PLUGINS                                   #
@@ -236,6 +243,13 @@ var serviceInputHeader = `
 
 ###############################################################################
 #                            SERVICE INPUT PLUGINS                            #
+###############################################################################
+`
+
+var webhookInputHeader = `
+
+###############################################################################
+#                            WEBHOOK INPUT PLUGINS                            #
 ###############################################################################
 `
 
@@ -285,10 +299,15 @@ func printFilteredInputs(inputFilters []string, commented bool) {
 	}
 	sort.Strings(pnames)
 
-	// cache service inputs to print them at the end
-	servInputs := make(map[string]telegraf.ServiceInput)
+	// cache service inputs to print them after inputs
+	servInputs := make(map[string]telegraf.Input)
 	// for alphabetical looping:
 	servInputNames := []string{}
+
+	// cache service inputs to print them after at the end
+	webhookInputs := make(map[string]telegraf.Input)
+	// for alphabetical looping:
+	webhookInputNames := []string{}
 
 	// Print Inputs
 	for _, pname := range pnames {
@@ -300,19 +319,31 @@ func printFilteredInputs(inputFilters []string, commented bool) {
 			servInputs[pname] = p
 			servInputNames = append(servInputNames, pname)
 			continue
+		case telegraf.WebhookInput:
+			webhookInputs[pname] = p
+			webhookInputNames = append(webhookInputNames, pname)
+			continue
 		}
 
 		printConfig(pname, input, "inputs", commented)
 	}
 
-	// Print Service Inputs
-	if len(servInputs) == 0 {
+	// Print service input
+	printOtherInputs(serviceInputHeader, servInputs, servInputNames, commented)
+
+	// Print webhook input
+	printOtherInputs(webhookInputHeader, webhookInputs, webhookInputNames, commented)
+}
+
+func printOtherInputs(header string, inputs map[string]telegraf.Input, names []string, commented bool) {
+	if len(names) == 0 {
 		return
 	}
-	sort.Strings(servInputNames)
-	fmt.Printf(serviceInputHeader)
-	for _, name := range servInputNames {
-		printConfig(name, servInputs[name], "inputs", commented)
+
+	sort.Strings(names)
+	fmt.Printf(header)
+	for _, name := range names {
+		printConfig(name, inputs[name], "inputs", commented)
 	}
 }